+++ app/src/main/java/com/example/crosshelper/ui/pixeling/PixelingFragment.java
@@ -1,4 +1,4 @@
-package com.example.crosshelper.ui.dashboard;
+package com.example.crosshelper.ui.pixeling;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -10,22 +10,22 @@ import androidx.annotation.NonNull;
 import androidx.fragment.app.Fragment;
 import androidx.lifecycle.ViewModelProvider;
 
-import com.example.crosshelper.databinding.FragmentDashboardBinding;
+import com.example.crosshelper.databinding.FragmentPixelingBinding;
 
-public class DashboardFragment extends Fragment {
+public class PixelingFragment extends Fragment {
 
-    private FragmentDashboardBinding binding;
+    private FragmentPixelingBinding binding;
 
     public View onCreateView(@NonNull LayoutInflater inflater,
                              ViewGroup container, Bundle savedInstanceState) {
-        DashboardViewModel dashboardViewModel =
-                new ViewModelProvider(this).get(DashboardViewModel.class);
+        PixelingViewModel pixelingViewModel =
+                new ViewModelProvider(this).get(PixelingViewModel.class);
 
-        binding = FragmentDashboardBinding.inflate(inflater, container, false);
+        binding = FragmentPixelingBinding.inflate(inflater, container, false);
         View root = binding.getRoot();
 
-        final TextView textView = binding.textDashboard;
-        dashboardViewModel.getText().observe(getViewLifecycleOwner(), textView::setText);
+        final TextView textView = binding.textNotifications;
+        pixelingViewModel.getText().observe(getViewLifecycleOwner(), textView::setText);
         return root;
     }
