+++ app/src/main/res/layout/fragment_pixeling.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.notifications.NotificationsFragment">
+    tools:context=".ui.pixeling.PixelingFragment">
 
     <TextView
         android:id="@+id/text_notifications"
