+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -13,13 +13,13 @@
 
     <fragment
         android:id="@+id/navigation_dashboard"
-        android:name="com.example.crosshelper.ui.dashboard.DashboardFragment"
+        android:name="com.example.crosshelper.ui.dashboard.FavoritesFragment"
         android:label="@string/title_favorites"
-        tools:layout="@layout/fragment_dashboard" />
+        tools:layout="@layout/fragment_favorites" />
 
     <fragment
         android:id="@+id/navigation_notifications"
-        android:name="com.example.crosshelper.ui.notifications.NotificationsFragment"
+        android:name="com.example.crosshelper.ui.pixeling.PixelingFragment"
         android:label="@string/title_pixeling"
-        tools:layout="@layout/fragment_notifications" />
+        tools:layout="@layout/fragment_pixeling" />
 </navigation>
